refactor(build-utils): js->ts
