+++ build-utils/tsconfig.json
@@ -10,6 +10,7 @@
           "es2016",
           "es2015",
           "es5",
+          "dom"
       ],
       "strict": true,
       "noLib": false,
